pyproject.toml
[project]
name = "styling-guide-demo"
version = "0.8.0"
description = "A proof-reader application using Azure OpenAI Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.99.9",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.32.4",
    "python-docx>=1.2.0",
    "lxml>=4.9.0",
    "gunicorn>=23.0.0",
    "python-Levenshtein>=0.27.1",
    "opencc==1.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "static/**/*",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
